+++ components/Contact/ContactForm.tsx
@@ -72,7 +72,7 @@ export default function ContactForm() {
       onSubmit={onSubmit}
     >
       {({ errors, touched }) => (
-        <Form role="form" id="contact-form" className="flex flex-col font-semibold font-montserrat gap-y-2 text-base md:text-lg">
+        <Form role="form" id="contact-form" className="flex flex-col font-semibold font-montserrat gap-y-4 text-base md:text-lg">
           <label htmlFor="fullName">Full Name:</label>
           <Field
             className={`border ${errors.fullName && touched.fullName ? 'border-black shadow-xl' : 'border-black'}
