+++ utils/window.ts
@@ -1,0 +1,9 @@
+export const openDocumentationPage = () => {
+  const url = "https://docs.specollective.org/introduction/";
+  window.open(url, '_blank');
+}
+
+export const openDonatePage = () => {
+  const url = "https://opencollective.com/spec";
+  window.open(url, '_blank');
+}
